(feature/SupportMinGw64GccBuild) Supported masp

+++ Programs/Sources/Libraries/masp/Basics/Window/main.cpp
@@ -1,3 +1,27 @@
+/*
+ * The MIT License (MIT)
+ *
+ * Copyright (c) 2015-2016 Ryosuke Kanata
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
+
 #include <stdio.h>
 #include <stdint.h>
 
@@ -20,9 +44,9 @@ static void test_window(void)
     masp::window::Kaiser(k, 3.0);
     masp::window::Flattop(f);
     printf(",Hanning,Hamming,Blackman,BlackmanHarris,Nuttall,Kaiser,Flattop,Rectangular\n");
-    for (uint i = 0; i < N; ++i)
+    for (size_t i = 0; i < N; ++i)
     {
-        printf("%d,%f,%f,%f,%f,%f,%f,%f,%f\n", i, han[i], ham[i], b[i], bh[i], n[i], k[i], f[i], s[i]);
+        printf("%d,%f,%f,%f,%f,%f,%f,%f,%f\n", static_cast<int>(i), han[i], ham[i], b[i], bh[i], n[i], k[i], f[i], s[i]);
     }
     N = 25;
     s.Resize(N);
@@ -44,9 +68,9 @@ static void test_window(void)
     masp::window::Flattop(f);
     printf("\n");
     printf(",Hanning,Hamming,Blackman,BlackmanHarris,Nuttall,Kaiser,Flattop,Rectangular\n");
-    for (uint i = 0; i < N; ++i)
+    for (size_t i = 0; i < N; ++i)
     {
-        printf("%d,%f,%f,%f,%f,%f,%f,%f,%f\n", i, han[i], ham[i], b[i], bh[i], n[i], k[i], f[i], s[i]);
+        printf("%d,%f,%f,%f,%f,%f,%f,%f,%f\n", static_cast<int>(i), han[i], ham[i], b[i], bh[i], n[i], k[i], f[i], s[i]);
     }
     return ;
 }
@@ -59,7 +83,7 @@ static void test_tapps(void)
     //mfio::Csv csv("rectangular.csv");
     mfio::Csv csv("sinc.csv");
 
-    for (uint i = 0; i < sizeof(Ms)/sizeof(int); ++i )
+    for (size_t i = 0; i < sizeof(Ms)/sizeof(int); ++i )
     {
         const double fe = 0.25;
         const int N = Ms[i] + 1;
